Add next-game OG image for the games page
/:teamSlug/games/og.png renders the team's next game (opponent, date,
location, cancelled tag) in the same frame as the leaderboard image.
The shared frame moved to app/lib/og-image.server.tsx so both OG routes
stay consistent.
